+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- [workflow]: All packages now have a workflow.
+- [workflow]: All packages now have a workflow on GitHub.
 - [frontend]: Frontend got updated to 0.8.4.
 
 ## [1.2.0] - 2024-03-23